shm: make the shared-heap freelist ABA-safe via tagged-pointer head
The size-bucketed freelist added in 07bc74c42e59 used a single-word
head and a plain CAS, with the design note explicitly accepting the
ABA window on the grounds that the only consequence would be a missed
recycling opportunity.  That assessment was wrong.  A child terminated
this evening with:

  Program terminated with signal SIGSEGV.
  #0  __list_del (prev=0x0, next=0x0) at include/list.h:47
  #1  list_del (entry=0x7fa8d5786058)
  #2  __destroy_object (obj=0x7fa8d5786058, scope=OBJ_GLOBAL,
                        type=OBJ_MMAP_ANON, ...) at objects.c:572

faulting at offset 0x8 — the offset of `prev` within a list_head whose
`next` was NULL.  The list_head was zero because freelist_push() zeros
every slot on free, and the slot was on the destruction list because
two callers had been handed the same slot by the racy popper.

The mechanism is the textbook ABA on a lock-free stack.  freelist_pop()
loads old_head=X, dereferences X to read next=*X, then CASes head from
X to next.  If, between the load and the CAS, another thread pops X,
pops *X, and pushes X back (now linking to a different successor Z),
the CAS still observes head==X and succeeds — but it installs the
stale `next` we read earlier, which now points at a slot that has
already been handed back into active use by the interleaved popper.
Both callers then initialise the same slot, both add it to a (possibly
different) head->list, and the survivor's iteration walks a chain that
the loser has already zeroed via the poison-on-free path.  The crash
above is what that looks like at exit time.

The fix is the standard tagged-pointer mitigation.  The freelist head
type widens from `_Atomic uintptr_t` to `_Atomic uint64_t` and is
treated as a packed (version, ptr) tuple: bits 0-47 hold the canonical
user-space pointer (bit 47 is always 0 for user-space addresses on
x86_64, so no sign extension is required on read) and bits 48-63 hold
a monotonic 16-bit version counter that is incremented on every
successful push and pop.  CAS now compares the full 64-bit tuple, so
the A→B→A sequence above leaves the head as (X, ver+2) rather than
(X, ver) and the racer's CAS fails on the version mismatch instead of
silently installing a stale link.  The slot's stored next-pointer is
just the bare address — the version lives only in the head — so the
existing slot-zeroing on push is unchanged.

Why tagged pointer rather than a 128-bit DWCAS via cmpxchg16b: the
tagged variant is a drop-in replacement that needs only a single
64-bit CAS (already used everywhere in the file), needs no new
compiler flags, and needs no libatomic.  The 16-bit version is finite
in principle — a perfectly-timed sequence of exactly 65536 push/pop
pairs in the interval between a victim's load and its CAS would wrap
the version back to its original value and re-expose the race — but
for a process-bounded fuzzer with sub-microsecond critical sections
the probability is astronomically small.  If the race ever proves
observable in practice the head can be widened to a true 128-bit
(version, ptr) tuple and switched to a cmpxchg16b-based DWCAS without
any caller change.

The public API (alloc_shared_obj, free_shared_obj, alloc_shared_str,
free_shared_str, alloc_shared_strdup) is unchanged, as is the slot
zeroing on free, the bump-allocator fallback for above-bucket sizes,
and the RELAXED ordering on the CAS itself (the caller still publishes
the resulting object via add_object()'s RELEASE store, which remains
the actual synchronisation point for consumers).

+++ include/shm.h
@@ -108,16 +108,19 @@ struct shm_s {
 	 * Per-bucket freelist heads for the shared obj and str heaps.
 	 * NUM_SHM_FREELIST_BUCKETS fixed-size slots (8..1024 bytes, powers of
 	 * two); allocations above 1024 bytes bypass the freelist and use the
-	 * bump allocator directly.  Each head is a uintptr_t storing the
-	 * address of the most-recently-freed slot in that bucket (0 = empty).
-	 * The link to the next free slot is stored in the slot's own first
-	 * sizeof(uintptr_t) bytes (safe because the slot is, by definition,
-	 * not live when the link is written).  Manipulated by lock-free CAS in
-	 * freelist_push/pop in utils.c.
+	 * bump allocator directly.  Each head is a 64-bit tagged pointer:
+	 * the low 48 bits hold the address of the most-recently-freed slot
+	 * (0 = empty list) and the high 16 bits hold a monotonic version
+	 * counter that defeats the ABA race in freelist_pop (see the long
+	 * comment above the freelist primitives in utils.c).  The link to the
+	 * next free slot is stored in the slot's own first sizeof(uintptr_t)
+	 * bytes (safe because the slot is, by definition, not live when the
+	 * link is written).  Manipulated by lock-free CAS in freelist_push/pop
+	 * in utils.c.
 	 */
 #define NUM_SHM_FREELIST_BUCKETS 8
-	_Atomic uintptr_t shared_obj_freelist[NUM_SHM_FREELIST_BUCKETS];
-	_Atomic uintptr_t shared_str_freelist[NUM_SHM_FREELIST_BUCKETS];
+	_Atomic uint64_t shared_obj_freelist[NUM_SHM_FREELIST_BUCKETS];
+	_Atomic uint64_t shared_str_freelist[NUM_SHM_FREELIST_BUCKETS];
 
 	/* various flags. */
 	enum exit_reasons exit_reason;
